safari-600.7-branch/LayoutTests/media/media-source/media-source-end-of-stream-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(source.endOfStream("network"))
EVENT(error)
EXPECTED (video.error.code == '4') OK
RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(source.endOfStream("decode"))
EVENT(error)
EXPECTED (video.error.code == '4') OK
RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(source.endOfStream("network"))
EVENT(error)
EXPECTED (video.error.code == '2') OK
RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(source.endOfStream("decode"))
EVENT(error)
EXPECTED (video.error.code == '3') OK
RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(source.endOfStream())
EVENT(suspend)
EXPECTED (video.networkState == '1') OK
END OF TEST
